Extract entity narration method as util method
